ref(metrics): Move MRI into relay-base-schema module
